Fixed the Submit button, maintaining the website static!
Essentially I learned the difference between a Static and Dynamic
website.

A Static website simply sources the html, css and image files to the
user and lets the browser handle rendering them.

However, the php code that was going to be run when the user pressed
the submit button requires an actual server to run it, and hence this
requires an AWS EC2 instance, with Load Balancing, and all sorts of
other services, to get involved, hence it gets a lot more complicated
quickly.

Honestly it doesn’t seem like anything incredibly complicated, but its
time consuming and its not relevant to the actual development of the
VReel app. If I need to build a more involved website in the future
then I will hire a contractor programmer to help teach me how to do
this, so that I don’t waste as much time next time.

My fix to get around this is that I POST to
“http://formspree.io/hello@vreel.io”, which sends hello@vreel.io the
contents of the form. Its a good middle ground, as it doesn’t require
me to make the website into a dynamic website (it can stay as a static
website), yet my submit button still achieves the desired result of
sending me the details of the form (ie. the person’s email address)!

This will lead to a certain amount of work in the long-run however, as
I will  need to compile all the email addresses we’ve received, but its
better than trying to make this website be dynamically hosted!