Created a Jobs page
Created a “jobs.html” page in order to be able to point people in that
direction. Its just a hack at my main index.html page and shows that we
need a Designer and Developer.

Also moved the feedback page into a Feedback folder and called it
index.html, so that now you only need to type vreel.io/feedback, as
opposed to feedback.html which I think is ugly!